+++ R/use_skill_create_issue.R
@@ -13,7 +13,7 @@
 use_skill_create_issue <- function(
   target_dir = ".github",
   use_skills_subdir = TRUE,
-  overwrite = TRUE,
+  overwrite = FALSE,
   open = rlang::is_interactive(),
   gh_token = gh::gh_token()
 ) {